♻️ Refactoring out file filter and tag filter test to file list view
