+++ src/documents/views/__tests__/StudyFilesListView.test.js
@@ -31,6 +31,131 @@ it('renders correctly', async () => {
   expect(tree.container).toMatchSnapshot();
 });
 
+it('renders with files pagination', async () => {
+  const tree = render(
+    <MockedProvider mocks={mocks}>
+      <MemoryRouter initialEntries={['/study/SD_X6E5KWM4/documents']}>
+        <Routes />
+      </MemoryRouter>
+    </MockedProvider>,
+  );
+  await wait();
+
+  // Sort files by create date by clicking on the file sorting dropdown
+  const sortDropdown = tree.getByText(/Date option/i);
+  fireEvent.click(sortDropdown);
+  await wait();
+  const sortByCreateDate = tree.getByText(/Modified date/i);
+  fireEvent.click(sortByCreateDate);
+  await wait(10);
+  expect(tree.container).toMatchSnapshot();
+
+  // 12 files in total, page 1 contains 10
+  const lisPage1 = tree.queryAllByTestId('file-item');
+  expect(lisPage1.length).toBe(10);
+
+  // Click on the next button to go to next page
+  let button1 = tree.getByLabelText('Next page');
+  button1.click();
+  await wait();
+  expect(tree.container).toMatchSnapshot();
+
+  // 12 files in total, 10 files per page, page 2 should have 2 files
+  const lisPage2 = tree.queryAllByTestId('file-item');
+  expect(lisPage2.length).toBe(2);
+
+  // Click on the previous button to go to previous page
+  let button2 = tree.getByLabelText('Previous page');
+  button2.click();
+  await wait();
+
+  // Filter files by file type by clicking on the document type dropdown
+  const filterDropdown = tree.getByText(/Document type/i);
+  fireEvent.click(filterDropdown);
+  await wait();
+  const filterByFileType = tree.queryAllByText(/Biospecimen Manifest/i)[1];
+  fireEvent.click(filterByFileType);
+  await wait();
+
+  expect(tree.container).toMatchSnapshot();
+  const lisPage3 = tree.queryAllByTestId('file-item');
+  expect(lisPage3.length).toBe(6);
+});
+
+it('renders with files with tag filter', async () => {
+  const tree = render(
+    <MockedProvider mocks={mocks}>
+      <MemoryRouter initialEntries={['/study/SD_8WX8QQ06']}>
+        <StudyFilesListView
+          history={[]}
+          match={{params: {kfId: 'SD_8WX8QQ06'}}}
+        />
+      </MemoryRouter>
+    </MockedProvider>,
+  );
+  await wait();
+  expect(tree.container).toMatchSnapshot();
+
+  // Sort files by create date by clicking on the file sorting dropdown
+  act(() => {
+    fireEvent.click(tree.getByText(/Date option/i));
+  });
+  act(() => {
+    fireEvent.click(tree.getByText(/Modified date/i));
+  });
+  await wait();
+  expect(tree.container).toMatchSnapshot();
+
+  const lisPage1 = tree.queryAllByTestId('file-item');
+  expect(lisPage1.length).toBe(2);
+
+  act(() => {
+    fireEvent.click(tree.getAllByText(/Tag/i)[0]);
+  });
+  expect(tree.container).toMatchSnapshot();
+  act(() => {
+    fireEvent.click(tree.getAllByText(/dbGaP/i)[0]);
+  });
+  await wait();
+  expect(tree.container).toMatchSnapshot();
+  const filteredPage = tree.queryAllByTestId('file-item');
+  expect(filteredPage.length).toBe(1);
+});
+
+it('renders with files with more tag toggle', async () => {
+  const tree = render(
+    <MockedProvider mocks={mocks}>
+      <MemoryRouter initialEntries={['/study/SD_8WX8QQ06']}>
+        <StudyFilesListView
+          history={[]}
+          match={{params: {kfId: 'SD_8WX8QQ06'}}}
+        />
+      </MemoryRouter>
+    </MockedProvider>,
+  );
+  await wait();
+  expect(tree.container).toMatchSnapshot();
+
+  // Sort files by create date by clicking on the file sorting dropdown
+  act(() => {
+    fireEvent.click(tree.getByText(/Modified date/i));
+  });
+  await wait();
+  expect(tree.container).toMatchSnapshot();
+  expect(tree.queryAllByText('hidden tag').length).toBe(1);
+  act(() => {
+    fireEvent.click(tree.getByText(/1 more/i));
+  });
+  expect(tree.container).toMatchSnapshot();
+  expect(tree.queryAllByText('hidden tag').length).toBe(2);
+  act(() => {
+    fireEvent.click(tree.getByText(/show less/i));
+  });
+  await wait();
+  expect(tree.container).toMatchSnapshot();
+  expect(tree.queryAllByText('hidden tag').length).toBe(1);
+});
+
 it('deletes a file correctly', async () => {
   const tree = render(
     <MockedProvider mocks={[mocks[1], mocks[3], mocks[8], mocks[26]]}>
